transaction entity for data layer

+++ features/payment/src/main/java/com/example/payment/domain/entity/TransactionEntity.kt
@@ -6,5 +6,5 @@ data class TransactionEntity(
     val date: Date,
     val foundationId: Long,
     val foundationName: String,
-    val sum: Int,
+    val sum: Long,
 )
